Refactor - extract products repo and rules repo
Renamed pricing rule to product repository
