+++ ruby/lib/products_repository.rb
@@ -1,0 +1,12 @@
+class ProductsRepository
+  RULES = [
+    { code: 'ult_small',  name: 'Unlimited 1GB', price: 24.9 },
+    { code: 'ult_medium', name: 'Unlimited 2GB', price: 29.9 },
+    { code: 'ult_large',  name: 'Unlimited 5GB', price: 44.9 },
+    { code: '1gb',        name: '1 GB Data-pack', price: 9.90 }
+  ]
+
+  def self.find_by_product_code(product_code)
+    RULES.select { |i| i[:code] == product_code }.first
+  end
+end
